+++ lib/unionlabs/fuzz/fuzz_targets/duration/roundtrip.rs
@@ -1,7 +1,7 @@
 #![no_main]
 
 use libfuzzer_sys::fuzz_target;
-use unionlabs::{ibc::google::protobuf::duration::Duration, test_utils::*};
+use unionlabs::{google::protobuf::duration::Duration, test_utils::*};
 
 fuzz_target!(|data: Duration| {
     assert_proto_roundtrip(&data);
