+++ lib/unionlabs/src/ibc.rs
@@ -1,4 +1,3 @@
 pub mod applications;
 pub mod core;
-pub mod google;
 pub mod lightclients;
