+++ tools/todo-comment.nix
@@ -6,6 +6,7 @@
         name = "tdc";
         runtimeInputs = with pkgs; [ gh coreutils ];
         text = ''
+          gh auth login
           TODO_ID=$(gh issue create -t "$1" -b "_Generated by the devShell's \`tdc\` cli_" --label "todo comment" -a "@me" | tail -n 1 | sed 's|https://github.com/\(.*\)/issues/\(.*\)|\1#\2|');
           echo "TODO($TODO_ID): $1"
         '';
